Eliminate duplicated code
